rewrite(web): slice 9s.1 — round-1 review fixes
Four reviewers (architecture, correctness, code-quality,
vision) approved with notes. This commit addresses the
must/should-fix items.

Architecture MED + boundary fix: move `bootstrap_default()`
to `impl TileLayout` in the shared crate
- Was a free fn in `crates/web/src/tile/layout.rs` that
  constructed only `katulong_shared::wire` types — pure
  layer-boundary violation. The future persistence slice
  may want to mint the same default server-side; either
  the function lives in shared (one source of truth) or
  the server gets a divergent reimplementation. Fixed it
  here, not later.
- Deleted `crates/web/src/tile/layout.rs` (the only thing
  in it was `bootstrap_default`). Future dispatch helpers
  (`add_tile`, `remove_tile`, etc.) can land in
  `mod.rs` or be split out at that point.

Code-Quality MED: `LayoutState` is a named-field struct
- Was `pub struct LayoutState(pub RwSignal<TileLayout>)`
  (tuple struct). `AuthState` next to it uses the named-
  field form `{ phase, set_phase }`. Future slices that
  add associated values to `LayoutState` (a `set_layout`
  reset path, an undo stack, dispatch helpers as methods)
  would have had to either add positional fields (each one
  touching every consumer) or refactor to named-field
  later. Doing it now while the consumer count is one is
  cheap.

Code-Quality MED: design doc snippet matches live code
- The component contract example showed
  `expect_context::<RwSignal<TileLayout>>()` but the actual
  code provides/consumes the `LayoutState` wrapper. A
  future tile author copying the doc snippet would have
  hit a runtime `expect_context` panic. Updated the
  snippet.

Code-Quality MED: collapse the dead branch in TerminalTile
- Both arms of `if attached` rendered the same string
  with a one-word difference; the comment said the
  variable existed "to document the shape." Per
  CLAUDE.md, comments document the WHY; a dead branch
  with a comment doesn't earn its keep. Replaced with a
  single rendering and a `let _ = session_id;` so a
  future reader still finds the prop is bound.

Correctness HIGH/MED + Code-Quality HIGH: serde-shape
guard comments
- Added a doc-comment block on `TileKind` warning future
  authors NOT to add `#[serde(skip_serializing_if =
  "Option::is_none")]` to `Option`-typed variant fields.
  The `tile_kind_terminal_serializes_with_session_id`
  test pins `session_id: None` as `"session_id": null`
  (present-but-null); skip_serializing_if would silently
  change the wire shape to omit the field, and consumers
  that distinguish absent from null would break.
- Added a doc-comment block on `TileDescriptor` flagging
  the `#[serde(flatten)]` + internally-tagged enum
  combination as version-sensitive across the serde
  ecosystem. The round-trip tests
  (`tile_descriptor_flattens_kind_at_root`,
  `tile_layout_round_trips_with_descriptors`) are the
  regression guards if a serde version bump regresses
  this.

Architecture HIGH: Cluster recursion decision recorded
- The deferral previously said "decide later"; the
  reviewer correctly flagged that the choice has a
  cascading invariant consequence. The doc now records
  two concrete options:
  - A: cluster children stored in `tiles`, hidden from
    `order` (forces invariant change to "subset" or to
    split `visible_order` from `all_tiles`)
  - B: layout becomes tree-shaped (cleaner invariant,
    bigger reshape)
  The decision still defers, but a future-cluster-slice
  author now has the two options + their consequences in
  one place, not just "decide later."

Architecture MED: logout-clear obligation comment
- Added a comment at the seed effect in `<App/>` and a
  bullet in the design doc's open questions: when a
  future logout slice writes
  `set_phase.set(SignedOut)`, it MUST also write
  `layout.set(TileLayout::empty())`. Without that, the
  next sign-in's seed effect sees non-empty tiles and
  skips the seed; the user would inherit stale tiles
  (with stale session_ids) from the previous session.

Architecture LOW + Vision LOW: invariant obligations and
HashMap rationale documented at the field
- `TileLayout`'s doc-comment now spells out (a) the
  permutation invariant for `order`, (b) the
  focused_id-iff-non-empty invariant, (c) the dangerous
  helpers (`remove_tile` advancing focus, clearing
  focused_id when last tile removed), and (d) the
  rationale for HashMap over Vec-of-pairs (keyed
  mutation pattern, not asymptotics).

Vision MED: Status tile is layout debt
- `bootstrap_default`'s doc-comment notes the Status tile
  is unreachable in this slice (host renders only the
  focused tile; no tab bar yet). The persistence slice,
  if it lands before the tab-bar slice, must treat the
  orphaned Status tile as expected layout content, not
  garbage to migrate out.

Code-Quality LOW: bootstrap IDs are stable, not UUIDs
- The `TileId` doc says "UUID-ish" but `bootstrap_default`
  uses `"default-terminal"`/`"default-status"` —
  stable well-known strings. Updated the `TileId` doc to
  acknowledge the bootstrap exception explicitly.

Code-Quality LOW: dropped slice version numbers from
inline comments. Architectural rationale stays.

Architecture / Vision: topic broker injection-via-context
- Doc now explicitly states tiles consume cross-cutting
  context (ConnectionStatus, AuthState, future
  TopicBroker) via `expect_context`, NOT via TileKind
  variant props. Subscription metadata (which topic
  this tile is interested in) belongs in the descriptor;
  runtime context is provided by the host's surroundings.

Defensible defers (one false positive, two intentional)
- Code-Quality reviewer flagged the
  `tile_kind_terminal_serializes_with_session_id` test
  as missing. The test exists at
  `crates/shared/tests/wire_round_trip.rs:259-282` and
  passes (`cargo test --release` confirms). Treating as
  reviewer false-positive.
- TileFrame shared-component extraction (Code-Quality
  INFO): premature at 2 tiles; will revisit at 5+.
- `pub(crate)` on TileLayout fields (Vision LOW):
  defensible to do now, but the wins (compile-time
  enforcement of dispatch-only mutation) are conditional
  on the dispatch helpers existing — and they don't yet.
  Defer to the dispatch-helpers slice, which has both
  the helpers and the visibility tightening as paired
  changes.

Verified
- `cargo build --release` clean.
- `cargo test --release -p katulong-shared` green: 12
  tests (bootstrap_default move didn't change the test
  count — the helper isn't tested directly; layouts it
  produces are covered by `tile_layout_round_trips_with_descriptors`).
- `trunk build --release` clean; bundle 591 KB → 592 KB
  (within noise; the LayoutState named-field rename and
  the bootstrap_default move don't materially affect
  codegen).
- Full Rust e2e (13 tests) green. The behavior under
  test (signed-in user lands on a "Signed in" view) is
  unchanged — the doc + comment + structural fixes
  preserve user-visible behavior.
